health/requirements.txt
connexion[swagger-ui]==2.14.2
PyYAML
requests
apscheduler